Refactor: 메인페이지 구조 변경
SPA 를 따르기 위해 메인 구조 변경
