+++ src/app/component/contents/Thumbnail.tsx
