+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import '../globals.css'
+import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import { GNB } from './component/GNB'
